refactor: resolving imports so that the project continues running
