+++ libs/atqr/domain/src/lib/EmailService/index.ts
@@ -1,4 +1,3 @@
-export * from './email';
 export * from './interfaces';
 export * from './email.service';
 export * from '../EmailService/templates/';
