+++ libs/atqr/domain/src/lib/valueObjects/emailAddress.ts
@@ -1,9 +1,8 @@
 import { ValueObject } from "../../utils/valueObject";
 
 export class EmailAddress extends ValueObject<EmailAddress> {
-  public value: string;
   // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  constructor(value: string) {
+  constructor(public readonly value: string) {
     super(EmailAddress, ['value'])
   }
