+++ libs/atqr/domain/src/lib/valueObjects/index.ts
@@ -1,5 +1,4 @@
-export * from './contact' 
+export * from './contact'
 export * from './email'
+export * from './emailAddress'
 export * from './invitee'
-export * from './enrollment'
-export * from './officiation'
